feat(containers): move drink&navi content to content dir
