+++ src/containers/content/CoolNaviContent/styles/content.js
