+++ src/containers/content/HaveADrinkContent/Header/IndexStatus.js
