+++ src/containers/content/HaveADrinkContent/schema.js
