+++ src/containers/content/HaveADrinkContent/styles/header/index_status.js
